Fix graph size adaption when container resizes
