+++ frontend/package.json
@@ -26,6 +26,7 @@
     "@cypress/react": "^5.4.0",
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
+    "@react-hook/resize-observer": "^1.2.0",
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
@@ -65,12 +66,12 @@
     "eslint-plugin-react": "^7.23.2",
     "eslint-plugin-react-hooks": "^4.2.0",
     "prettier": "^2.2.1",
+    "run-script-os": "^1.1.6",
     "stylelint": "^13.13.1",
     "stylelint-config-airbnb": "^0.0.0",
     "stylelint-order": "^4.1.0",
     "stylelint-scss": "^3.19.0",
-    "stylelint-webpack-plugin": "^2.1.1",
-    "run-script-os": "^1.1.6"
+    "stylelint-webpack-plugin": "^2.1.1"
   },
   "browserslist": {
     "production": [
